creacion de sesion para la aplicacion
creacion del modulo de sesion usando la gema "devise"

+++ db/migrate/20150310134236_create_daw_calificacions.rb
@@ -1,7 +1,7 @@
 class CreateDawCalificacions < ActiveRecord::Migration
   def change
     create_table :daw_calificacions do |t|
-      t.integer :calif_notaobtenida
+      t.float :calif_notaobtenida
       t.boolean :calif_presento
       t.date :calif_fecha
       t.references :daw_actividad, index: true
